deal with hash randomization in tests
This is an unfortunate change, really.  Ideally, _CH here would return:

  all(subhashof(\%hash),superhashof(\%hash));

Or, even better:

  samehash(\%hash);

But:

a. somehash() does not exist (yet)
b. we only get the list of k/v in list context, so we needed to use a
   listmethods() test, and I don't think we can do something like:

   listmethods('methodname' => subhashof(...))

   ...when we get back a list of k/v, rather than a hashref.

I implemented this in a way that I can fix it if Test::Deep (or my
understanding of it) improves.